+++ CHANGELOG.md
@@ -4,6 +4,9 @@ GitLab Dependency Scanning follows versioning of GitLab (`MAJOR.MINOR` only) and
 
 These "stable" Docker images may be updated after release date, changes are added to the corresponding section bellow.
 
+## 11-4-stable
+- Fix dependency scanning ignoring the variable DEP_SCAN_DISABLE_REMOTE_CHECKS.
+
 ## 11-3-stable
 
 ## 11-2-stable